+++ Sources/MapboxMaps/Gestures/GestureHandlers/SingleTapGestureHandler.swift
@@ -3,9 +3,13 @@ import UIKit
 /// `SingleTapGestureHandler` manages a gesture recognizer looking for single tap touch events
 internal final class SingleTapGestureHandler: GestureHandler {
 
-    internal init(gestureRecognizer: UITapGestureRecognizer) {
+    private let cameraAnimationsManager: CameraAnimationsManagerProtocol
+
+    internal init(gestureRecognizer: UITapGestureRecognizer,
+                  cameraAnimationsManager: CameraAnimationsManagerProtocol) {
         gestureRecognizer.numberOfTapsRequired = 1
         gestureRecognizer.numberOfTouchesRequired = 1
+        self.cameraAnimationsManager = cameraAnimationsManager
         super.init(gestureRecognizer: gestureRecognizer)
         gestureRecognizer.addTarget(self, action: #selector(handleGesture(_:)))
     }
@@ -13,6 +17,7 @@ internal final class SingleTapGestureHandler: GestureHandler {
     @objc private func handleGesture(_ gestureRecognizer: UITapGestureRecognizer) {
         switch gestureRecognizer.state {
         case .recognized:
+            cameraAnimationsManager.cancelAnimations()
             delegate?.gestureBegan(for: .singleTap)
             delegate?.gestureEnded(for: .singleTap, willAnimate: false)
         default: